requirements.txt
requests>=2.25.1
beautifulsoup4>=4.9.3
pandas>=1.3.0
openai>=1.0.0
python-dotenv>=1.0.0